refactor(emitter): split control-flow recovery helpers
